requirements.txt
Django==1.6.7
42cc-pystyle
South==1.0.2
Pillow==3.3.1
